+++ src/frontend/components/DashboardVideoLiveConfigureButton/index.tsx
@@ -5,7 +5,7 @@ import { Redirect } from 'react-router-dom';
 import { initiateLive } from '../../data/sideEffects/initiateLive';
 import { useVideo } from '../../data/stores/useVideo';
 import { modelName } from '../../types/models';
-import { Video } from '../../types/tracks';
+import { LiveModeType, Video } from '../../types/tracks';
 import { Nullable } from '../../utils/types';
 import { DASHBOARD_ROUTE } from '../Dashboard/route';
 import { DashboardButton } from '../DashboardPaneButtons';
@@ -13,10 +13,15 @@ import { FULL_SCREEN_ERROR_ROUTE } from '../ErrorComponents/route';
 import { Loader } from '../Loader';
 
 const messages = defineMessages({
-  btnConfigureLive: {
+  raw: {
     defaultMessage: 'Configure a live streaming',
     description: 'Dashboard button to configure a live streaming',
-    id: 'components.Dashboard.DashboardPaneButtons.videos.btnStartLive',
+    id: 'components.Dashboard.DashboardPaneButtons.videos.raw',
+  },
+  jitsi: {
+    defaultMessage: 'Launch Jitsi LiveStream',
+    description: 'Dashboard button to launch jitsi livestream',
+    id: 'components.Dashboard.DashboardPaneButtons.videos.jitsi',
   },
 });
 
@@ -25,10 +30,12 @@ type configureLiveStatus = 'pending' | 'success' | 'error';
 /** Props shape for the DashboardVideoLiveConfigureButton component. */
 export interface DashboardVideoLiveConfigureButtonProps {
   video: Video;
+  type: LiveModeType;
 }
 
 export const DashboardVideoLiveConfigureButton = ({
   video,
+  type,
 }: DashboardVideoLiveConfigureButtonProps) => {
   const [status, setStatus] = useState<Nullable<configureLiveStatus>>(null);
   const { updateVideo } = useVideo((state) => ({
@@ -38,7 +45,7 @@ export const DashboardVideoLiveConfigureButton = ({
   const configureLive = async () => {
     setStatus('pending');
     try {
-      const updatedVideo = await initiateLive(video);
+      const updatedVideo = await initiateLive(video, type);
       updateVideo(updatedVideo);
       setStatus('success');
     } catch (error) {
@@ -60,7 +67,7 @@ export const DashboardVideoLiveConfigureButton = ({
 
       <DashboardButton
         onClick={configureLive}
-        label={<FormattedMessage {...messages.btnConfigureLive} />}
+        label={<FormattedMessage {...messages[type]} />}
       />
     </React.Fragment>
   );